+++ src/nodes/textures/TextureNode.ts
@@ -6,8 +6,8 @@ import sockets from '../../rete/sockets';
 export class TextureNode extends Classic.Node {
     width = 180;
     height = 270;
-    texture: null | Uint8ClampedArray = null;
-    alpha: null | Uint8ClampedArray = null;
+    texture: null | Int16Array = null;
+    alpha: null | Int16Array = null;
     constructor(label: string) {
       super(label);
       this.addControl('preview', new Preview());
@@ -19,7 +19,7 @@ export class TextureNode extends Classic.Node {
     }
     makeTexture() {
         //Assumes RGBA
-        this.texture=new Uint8ClampedArray(4 * globals.textureRes * globals.textureRes);
+        this.texture=new Int16Array(4 * globals.textureRes * globals.textureRes);
         for (var y = 0; y < globals.textureRes; y++) {
           for (var x = 0; x < globals.textureRes; x++){
                 var cell = (x + y * globals.textureRes) * 4;
@@ -38,9 +38,9 @@ export class TextureNode extends Classic.Node {
     }
     data (inputs: any) {
       if(!this.texture) return {};
-      this.alpha = inputs['alpha'] && inputs['alpha'][0] instanceof Uint8ClampedArray ? inputs['alpha'][0] : null;
+      this.alpha = inputs['alpha'] && inputs['alpha'][0] instanceof Int16Array ? inputs['alpha'][0] : null;
       this.makeTexture();
       (this.controls['preview'] as Preview).setTexture(this.texture);
-      return {value:<Uint8ClampedArray>this.texture};
+      return {value:<Int16Array>this.texture};
     }
   }
